add close opportunity feature v1 https://github.com/reallocate/reallocate/issues/66
display a button to close it that’s not wired up at all for everyone, even non-admins. DONE
display button only if owner of the project. DONE
Change the status to STATUS_CLOSED w/out saving a comment or anything. use same modal style b/c it’s easier. DONE
persist the final message to db. hmmm, ask. maybe it just gets emailed out. not persisted to db of updates? DONE
When Opp is closed, notify all users DONE
If Opp is closed, then hide the ‘Engage’ button DONE
If Opp is closed put a visual  notification that it is closed.  - TODO
If Opp is closed, allow it to re-opened? N/A
If Opp is closed, show updates stream? DONE
If Opp is closed include the final ‘why it’s closed’ comment? (if there is one persisted in the first place?) DONE
Setup admin email template - TODO
Setup user email template - TODO